Upgrade for Symfony 5
Allow to run the bundle in Symfony 3.4+, 4.3+ and 5.0+.

Closes #10.
