+++ composer.json
@@ -24,24 +24,24 @@
         }
     ],
     "require": {
-        "php": "^7.1",
+        "php": "^7.2",
         "ext-soap": "*",
         "ramsey/uuid": "^3.7",
-        "symfony/config": "^3.0|^4.0",
-        "symfony/dependency-injection": "^3.0|^4.0",
-        "symfony/event-dispatcher": "~3.4|~4.0",
-        "symfony/http-kernel": "^3.0|^4.0"
+        "symfony/config": "^3.4 || ^4.3 || ^5.0",
+        "symfony/dependency-injection": "^3.4 || ^4.3 || ^5.0",
+        "symfony/event-dispatcher": "^3.4 || ^4.3 || ^5.0",
+        "symfony/http-kernel": "^3.4 || ^4.3 || ^5.0"
     },
     "require-dev": {
         "gamez/psr-testlogger": "^2.0",
         "phpunit/phpunit": "^7",
         "squizlabs/php_codesniffer": "^2.3",
-        "symfony/browser-kit": "^3.0|^4.0",
-        "symfony/debug-bundle": "^3.0|^4.0",
-        "symfony/framework-bundle": "^3.3|^4.0",
-        "symfony/phpunit-bridge": "^3.3|^4.0",
-        "symfony/web-profiler-bundle": "^3.0|^4.0",
-        "symfony/yaml": "~3.4|~4.0"
+        "symfony/browser-kit": "^3.4 || ^4.3 || ^5.0",
+        "symfony/debug-bundle": "^3.4 || ^4.3 || ^5.0",
+        "symfony/framework-bundle": "^3.4 || ^4.3 || ^5.0",
+        "symfony/phpunit-bridge": "^3.4 || ^4.3 || ^5.0",
+        "symfony/web-profiler-bundle": "^3.4 || ^4.3 || ^5.0",
+        "symfony/yaml": "^3.4 || ^4.3 || ^5.0"
     },
     "autoload": {
         "psr-4": {
@@ -60,7 +60,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.1-dev"
+            "dev-master": "1.5-dev"
         }
     },
     "config": {
